+++ tests/test_primes.py
@@ -1,7 +1,9 @@
+from fractions import Fraction
 from math import isqrt
 from itertools import pairwise
+from typing import Mapping
 import pytest
-from xenterval._common import KNOWN_PRIMES, prime_index
+from xenterval._primes import KNOWN_PRIMES, prime_index, prime_faсtors
 
 
 def is_positive_prime(x: int) -> bool:
@@ -28,3 +30,14 @@ def test_prime_index() -> None:
         else:
             with pytest.raises(ValueError, match='not a prime'):
                 prime_index(x)
+
+@pytest.mark.parametrize(['ratio_str', 'factors'], [
+    ('1', {}),
+    ('81/1210', {2: -1, 3: 4, 5: -1, 11: -2}),
+    ('104060401', {101: 4}),
+    ('370069474229841/734815760', {541: 2, 149: 1, 53: 3, 3: 1, 19: 1,
+                                   2: -4, 5: -1, 7: -3, 439: -1, 61: -1}),
+])
+def test_prime_factors(ratio_str: str, factors: Mapping[int, int]) -> None:
+    actual_factors = prime_faсtors(Fraction(ratio_str))
+    assert actual_factors == factors