+++ tests/test_xenterval.py
@@ -1,5 +1,0 @@
-from xenterval import __version__
-
-
-def test_version():
-    assert __version__ == '0.1.2'